BUG: Fix bug with doc rebuilds being forced
