+++ doc/sphinxext/contrib_avatars.py
@@ -3,12 +3,14 @@
 import os
 from pathlib import Path
 
+from mne.utils import _replace_md5
+
 
 def generate_contrib_avatars(app, config):
     """Render a template webpage with avatars generated by JS and a GitHub API call."""
     root = Path(app.srcdir)
     infile = root / "sphinxext" / "_avatar_template.html"
-    outfile = root / "_templates" / "avatars.html"
+    outfile = root / "_templates" / "avatars.html.new"
     if os.getenv("MNE_ADD_CONTRIBUTOR_IMAGE", "false").lower() != "true":
         body = """\
 <p>Contributor avators will appear here in full doc builds. Set \
@@ -36,6 +38,7 @@ MNE_ADD_CONTRIBUTOR_IMAGE=true in your environment to generate it.</p>"""
         driver.quit()
     with open(outfile, "w") as fid:
         fid.write(body)
+    _replace_md5(str(outfile))
 
 
 def setup(app):
