+++ mne/utils/__init__.pyi
@@ -98,6 +98,7 @@ __all__ = [
     "_reg_pinv",
     "_reject_data_segments",
     "_repeated_svd",
+    "_replace_md5",
     "_require_version",
     "_resource_path",
     "_safe_input",
@@ -361,6 +362,7 @@ from .numerics import (
     _mask_to_onsets_offsets,
     _reg_pinv,
     _reject_data_segments,
+    _replace_md5,
     _ReuseCycle,
     _scaled_array,
     _stamp_to_dt,
